Sécurise les fonctionnalités delete en verifiant que la method est bien un POST

+++ src/com/crexos/main/utils/DeleteAuthor.java
@@ -11,8 +11,10 @@ public class DeleteAuthor extends AbstractAction
 	@Override
 	public Redirect executeAction(HttpServletRequest request)
 	{
-		DAOFactory.getInstance().getAuthorDAO().delete(Integer.parseInt(request.getParameter("author-id")));
-		
+		if(request.getMethod().equals("POST"))
+		{
+			DAOFactory.getInstance().getAuthorDAO().delete(Integer.parseInt(request.getParameter("author-id")));
+		}
 		return new Redirect(true, "authors");
 	}
